# TensorImpl: Lazily compute numel and contiguity when symbolic