--- a/c10/core/TensorImpl.h
+++ b/c10/core/TensorImpl.h
@@ -7,6 +7,7 @@
 #include <c10/core/Storage.h>
 #include <c10/core/SymBool.h>
 #include <c10/core/SymIntArrayRef.h>
+#include <c10/core/SymbolicShapeMeta.h>
 #include <c10/core/WrapDimMinimal.h>
 #include <c10/core/impl/PyObjectSlot.h>
 #include <c10/core/impl/SizesAndStrides.h>
@@ -218,19 +219,6 @@ struct C10_API BackendMeta : intrusive_ptr_target {
       const intrusive_ptr<BackendMeta>& ptr) const {
     return ptr;
   }
-};
-
-struct C10_API SymbolicShapeMeta {
-  SymDimVector sizes_ = {0};
-  SymDimVector strides_ = {1};
-  SymInt numel_ = 1;
-  SymInt storage_offset_ = 0;
-  SymBool is_contiguous_{true};
-  SymBool is_channels_last_contiguous_{false};
-  SymBool is_channels_last_3d_contiguous_{false};
-  SymBool is_channels_last_{false};
-  SymBool is_channels_last_3d_{false};
-  SymBool is_non_overlapping_and_dense_{true};
 };
 
 struct C10_API ExtraMeta {
@@ -710,7 +698,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
 
   c10::SymInt sym_numel_default() const {
     if (has_symbolic_sizes_strides_) {
-      return symbolic_shape_meta().numel_;
+      return symbolic_shape_meta().numel();
     } else {
       return c10::SymInt(SymInt::UNCHECKED, numel_);
     }
@@ -825,13 +813,14 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
   bool is_contiguous_default(at::MemoryFormat memory_format) const {
     if (has_symbolic_sizes_strides_) {
       if (memory_format == at::MemoryFormat::ChannelsLast) {
-        return symbolic_shape_meta().is_channels_last_contiguous_.guard_bool(
+        return symbolic_shape_meta().is_channels_last_contiguous().guard_bool(
             __FILE__, __LINE__);
       } else if (memory_format == at::MemoryFormat::ChannelsLast3d) {
-        return symbolic_shape_meta().is_channels_last_3d_contiguous_.guard_bool(
-            __FILE__, __LINE__);
+        return symbolic_shape_meta()
+            .is_channels_last_3d_contiguous()
+            .guard_bool(__FILE__, __LINE__);
       }
-      return symbolic_shape_meta().is_contiguous_.guard_bool(
+      return symbolic_shape_meta().is_contiguous().guard_bool(
           __FILE__, __LINE__);
     }
 
@@ -846,10 +835,10 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
   bool is_strides_like_default(at::MemoryFormat memory_format) const {
     if (has_symbolic_sizes_strides_) {
       if (memory_format == at::MemoryFormat::ChannelsLast) {
-        return symbolic_shape_meta().is_channels_last_.guard_bool(
+        return symbolic_shape_meta().is_channels_last().guard_bool(
             __FILE__, __LINE__);
       } else if (memory_format == at::MemoryFormat::ChannelsLast3d) {
-        return symbolic_shape_meta().is_channels_last_3d_.guard_bool(
+        return symbolic_shape_meta().is_channels_last_3d().guard_bool(
             __FILE__, __LINE__);
       } else {
         return false;
@@ -867,7 +856,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
 
   bool is_non_overlapping_and_dense_default() const {
     if (has_symbolic_sizes_strides_) {
-      return symbolic_shape_meta().is_non_overlapping_and_dense_.guard_bool(
+      return symbolic_shape_meta().is_non_overlapping_and_dense().guard_bool(
           __FILE__, __LINE__);
     } else {
       return is_non_overlapping_and_dense_;
@@ -2523,16 +2512,6 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
     return static_cast<int64_t>(n);
   }
 
-  SymInt compute_sym_numel() const {
-    TORCH_INTERNAL_ASSERT_DEBUG_ONLY(has_symbolic_sizes_strides_);
-    auto& sym_shape_meta{symbolic_shape_meta()};
-    SymInt numel = 1;
-    for (const auto& s : sym_shape_meta.sizes_) {
-      numel *= s;
-    }
-    return numel;
-  }
-
   /**
    * Compute whether or not a tensor is contiguous based on the sizes and
    * strides of a tensor.
@@ -2548,18 +2527,6 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
   bool compute_strides_like_channels_last_3d(identity<bool>) const;
 
   bool compute_non_overlapping_and_dense(identity<bool>) const;
-
-  SymBool compute_contiguous(identity<SymBool>) const;
-
-  SymBool compute_channels_last_contiguous_2d(identity<SymBool>) const;
-
-  SymBool compute_channels_last_contiguous_3d(identity<SymBool>) const;
-
-  SymBool compute_strides_like_channels_last_2d(identity<SymBool>) const;
-
-  SymBool compute_strides_like_channels_last_3d(identity<SymBool>) const;
-
-  SymBool compute_non_overlapping_and_dense(identity<SymBool>) const;
 
  protected:
   /**
@@ -2579,7 +2546,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
    */
   void refresh_numel() {
     if (has_symbolic_sizes_strides_) {
-      symbolic_shape_meta().numel_ = compute_sym_numel();
+      symbolic_shape_meta().refresh_numel();
     } else {
       numel_ = compute_numel();
     }
@@ -2595,7 +2562,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
     if (has_symbolic_sizes_strides_) {
       // NB: sym numel is done with symbolic integers, which handle overflow
       // checking
-      symbolic_shape_meta().numel_ = compute_sym_numel();
+      symbolic_shape_meta().refresh_numel();
     } else {
       numel_ = safe_compute_numel();
     }
@@ -2609,48 +2576,24 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
     is_contiguous_ = b;
   }
 
-  void _set_is_contiguous(identity<SymBool>, SymBool b) {
-    symbolic_shape_meta().is_contiguous_ = std::move(b);
-  }
-
   void _set_is_channels_last_contiguous(identity<bool>, bool b) {
     is_channels_last_contiguous_ = b;
-  }
-
-  void _set_is_channels_last_contiguous(identity<SymBool>, SymBool b) {
-    symbolic_shape_meta().is_channels_last_contiguous_ = std::move(b);
   }
 
   void _set_is_channels_last_3d_contiguous(identity<bool>, bool b) {
     is_channels_last_3d_contiguous_ = b;
   }
 
-  void _set_is_channels_last_3d_contiguous(identity<SymBool>, SymBool b) {
-    symbolic_shape_meta().is_channels_last_3d_contiguous_ = std::move(b);
-  }
-
   void _set_is_channels_last(identity<bool>, bool b) {
     is_channels_last_ = b;
-  }
-
-  void _set_is_channels_last(identity<SymBool>, SymBool b) {
-    symbolic_shape_meta().is_channels_last_ = std::move(b);
   }
 
   void _set_is_channels_last_3d(identity<bool>, bool b) {
     is_channels_last_3d_ = b;
   }
 
-  void _set_is_channels_last_3d(identity<SymBool>, SymBool b) {
-    symbolic_shape_meta().is_channels_last_3d_ = std::move(b);
-  }
-
   void _set_is_non_overlapping_and_dense(identity<bool>, bool b) {
     is_non_overlapping_and_dense_ = b;
-  }
-
-  void _set_is_non_overlapping_and_dense(identity<SymBool>, SymBool b) {
-    symbolic_shape_meta().is_non_overlapping_and_dense_ = std::move(b);
   }
 
   // These are little wrappers over the real compute_ functions that
@@ -2666,27 +2609,19 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
         compute_non_overlapping_and_dense(type_id);
   }
 
-  SymBool compute_is_non_overlapping_and_dense_dim4(identity<SymBool> type_id);
-
   bool compute_channels_last_contiguous_3d_dim5(identity<bool> type_id) {
     return !is_channels_last_contiguous_ &&
         compute_channels_last_contiguous_3d(type_id);
   }
-
-  SymBool compute_channels_last_contiguous_3d_dim5(identity<SymBool> type_id);
 
   bool compute_channels_last_2d_dim5(identity<bool> type_id) {
     return !is_channels_last_3d_contiguous_ &&
         compute_strides_like_channels_last_2d(type_id);
   }
 
-  SymBool compute_channels_last_2d_dim5(identity<SymBool> type_id);
-
   bool compute_channels_last_3d_dim5(identity<bool> type_id) {
     return !is_channels_last_ && compute_strides_like_channels_last_3d(type_id);
   }
-
-  SymBool compute_channels_last_3d_dim5(identity<SymBool> type_id);
 
   bool compute_is_non_overlapping_and_dense_dim5(identity<bool> type_id) {
     return is_contiguous_ || is_channels_last_contiguous_ ||
@@ -2694,14 +2629,9 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
         compute_non_overlapping_and_dense(type_id);
   }
 
-  SymBool compute_is_non_overlapping_and_dense_dim5(identity<SymBool> type_id);
-
   bool compute_is_non_overlapping_and_dense_anydim(identity<bool> type_id) {
     return is_contiguous_ || compute_non_overlapping_and_dense(type_id);
   }
-
-  SymBool compute_is_non_overlapping_and_dense_anydim(
-      identity<SymBool> type_id);
 
   template <typename T>
   void _refresh_contiguous() {
@@ -2761,7 +2691,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
    */
   void refresh_contiguous() {
     if (has_symbolic_sizes_strides_) {
-      _refresh_contiguous<SymBool>();
+      symbolic_shape_meta().refresh_contiguous();
     } else {
       _refresh_contiguous<bool>();
     }