@peterbell10 2023-11-07
Ran the benchmark suit and seeing around 10% compile time improvement for dynamic shapes.


| Suite       | Master  (s)            | This PR (s) | Speedup |
|-------------|-----------------------|---------|---------|
| TorchBench  | 85                    | 79     | 1.08    |
| HuggingFace | 122                   | 106     | 1.15    |
| TIMM Models | 147                   | 138     | 1.07    |
| [dynamic]   | 43                    | 29      | 1.48    |

Full results here:
https://hud.pytorch.org/benchmark/compilers?startTime=Tue%2C%2031%20Oct%202023%2011%3A40%3A21%20GMT&stopTime=Tue%2C%2007%20Nov%202023%2011%3A40%3A21%20GMT&granularity=hour&suite=torchbench&mode=training&dtype=amp&lBranch=gh/peterbell10/656/orig&lCommit=e236834d4186ce09934bee43ff28b1683d673757&rBranch=main&rCommit=2bc1378d7be563fa9b3050bb0e0fefd6e55a9e81