@ezyang 2023-11-07
This isn't thread safe.

It should be possible to make it thread safe with minimal impact, as we eventually will call into Python which will be GIL protected. However, I don't see a clear recipe for doing so. I suppose we could allocate an atomic per field (or maybe only a coarse grained atomic is good enough?)