@Flamefire 2024-01-10
While working on #117033 I found this PR and I'm a bit worried about using that mutex in `SymbolicShapeMeta` only in const methods. Is it impossible for 2 threads having references to the same instance and one accessing it non-const while the other reads it (i.e. as const)? That would have only the "other" lock it which means the data is still concurrently accessed.   
Second: The `available_.fetch_and(~numel_avail); numel_ = 1;`-pattern used in 2 places looks odd: If the `available`-set already means, `numel_` is invalid and needs to be recomputed on access, why set it? Can't now a race with `set_numel` happen, that does `numel_ = std::move(val); available_.fetch_or(numel_avail);`? I.e. one thread clears the flag, the other sets it and right after the first sets `numel_ =1` making the value be "available" but wrong.