@peterbell10 2024-01-10
> Is it impossible for 2 threads having references to the same instance and one accessing it non-const while the other reads it (i.e. as const)? 

That would indicate a bug as the mutex isn't intended to protect all concurrent access but only to provide the default expectation that concurrent const accesses won't race. This is a in line with the rest of `TensorImpl`.

> The `available_.fetch_and(~numel_avail); numel_ = 1;` -pattern used in 2 places looks odd: [...] Can't now a race with set_numel happen,

As before, `refresh_numel` is non-const so there is no expectation of threading guaruntees.

> If the `available`-set already means, `numel_` is invalid and needs to be recomputed on access, why set it?

numel may be a symbolic expression in python, so setting it releases that reference.